Refactor: More centralizing handlers and refactoring
